+++ js/nativeplot/types.ts
@@ -3,7 +3,7 @@ import type { LoadingStatus } from "@gradio/statustracker";
 
 export interface PlotData {
 	columns: string[];
-	data: [string | number][];
+	data: (string | number)[][];
 	datatypes: Record<string, "quantitative" | "temporal" | "nominal">;
 	mark: "line" | "point" | "bar";
 }