+++ js/uploadbutton/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @gradio/uploadbutton
 
+## 0.10.0
+
+### Features
+
+- [#13502](https://github.com/gradio-app/gradio/pull/13502) [`429faeb`](https://github.com/gradio-app/gradio/commit/429faeb643fb1afc1722c0f63fafa11603f2c87f) - Ensure every component dispatches a `change` event when its value changes.  Thanks @abidlabs!
+
+### Dependency updates
+
+- @gradio/client@2.2.2
+- @gradio/button@0.7.0
+
 ## 0.9.19
 
 ### Fixes